+++ src/xbeeRf.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <Arduino.h>
+
 #define HOST_ADDRESS 0x0003
 
 void setupXbee();
-void sendXbee(uint8_t *order, uint64_t length);
+void sendXbee(uint8_t *order, uint8_t length);
 int readXbee();
 
 // enum Direction {DIR_UP, DIR_DOWN, DIR_LEFT, DIR_RIGHT, DIR_NONE};
